water, production and energy plots for provinces
